pyproject.toml
[project]
name = "helix-engine"
version = "0.1.1"
description = "Evolutionary prompt optimization engine inspired by Mind Evolution"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core engine deps (no web server, no database)
    "httpx>=0.28.1",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.13.0",
    "jinja2>=3.1.6",
    "tenacity>=9.1.0",
    "openai>=2.8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
web = [
    "fastapi[standard]>=0.135.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.22.1",
    "asyncpg>=0.29.0",
    "cryptography>=42.0",
    "PyJWT>=2.8.0",
    "bcrypt>=4.0.0",
]

[dependency-groups]
dev = [
    "helix-engine[web]",
    "pillow>=12.1.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-m 'not e2e'"
markers = [
    "e2e: end-to-end tests requiring real API keys (deselect with '-m not e2e')",
    "benchmark: engine performance benchmarks (adaptive sampling, checkpoints)",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = [
    "E501",   # line length handled by formatter
    "B008",   # Depends() in function defaults is standard FastAPI pattern
    "SIM102", # nested if -- sometimes clearer than combined condition
    "SIM108", # ternary -- sometimes less readable than if/else
    "SIM105", # contextlib.suppress -- try/except/pass can be clearer with comments
    "RUF006", # fire-and-forget asyncio tasks are intentional in background processing
]
